refactor: domain 패키지의 하위 패키지 삭제
