+++ src/main/java/kitchenpos/application/OrderService.java
@@ -11,13 +11,13 @@ import kitchenpos.application.exception.MenuNotFoundException;
 import kitchenpos.application.exception.OrderNotFoundException;
 import kitchenpos.application.exception.OrderTableNotFoundException;
 import kitchenpos.domain.menu.Menu;
-import kitchenpos.domain.menu.repository.MenuRepository;
+import kitchenpos.domain.menu.MenuRepository;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItem;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.order.repository.OrderRepository;
+import kitchenpos.domain.order.OrderRepository;
 import kitchenpos.domain.ordertable.OrderTable;
-import kitchenpos.domain.ordertable.repository.OrderTableRepository;
+import kitchenpos.domain.ordertable.OrderTableRepository;
 import kitchenpos.ui.dto.request.CreateOrderLineItemRequest;
 import kitchenpos.ui.dto.request.CreateOrderRequest;
 import kitchenpos.ui.dto.request.UpdateOrderStatusRequest;
