+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -5,8 +5,8 @@ import java.util.stream.Collectors;
 import kitchenpos.application.dto.CreateTableGroupDto;
 import kitchenpos.application.exception.TableGroupNotFoundException;
 import kitchenpos.domain.tablegroup.TableGroup;
-import kitchenpos.domain.tablegroup.repository.TableGroupRepository;
-import kitchenpos.domain.tablegroup.service.GroupingTableService;
+import kitchenpos.domain.tablegroup.TableGroupRepository;
+import kitchenpos.domain.tablegroup.GroupingTableService;
 import kitchenpos.ui.dto.request.CreateOrderGroupOrderTableRequest;
 import kitchenpos.ui.dto.request.CreateTableGroupRequest;
 import org.springframework.stereotype.Service;
