+++ src/main/java/kitchenpos/domain/ordertable/OrderTableRepository.java
@@ -1,7 +1,6 @@
-package kitchenpos.domain.ordertable.repository;
+package kitchenpos.domain.ordertable;
 
 import java.util.List;
-import kitchenpos.domain.ordertable.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
