+++ src/main/java/kitchenpos/domain/product/ProductRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.domain.product.repository;
+package kitchenpos.domain.product;
 
-import kitchenpos.domain.product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
