+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -26,11 +26,11 @@ import kitchenpos.domain.ordertable.OrderTable;
 import kitchenpos.domain.product.Product;
 import kitchenpos.domain.exception.InvalidOrderLineItemException;
 import kitchenpos.domain.exception.InvalidOrderStatusException;
-import kitchenpos.domain.menugroup.repository.MenuGroupRepository;
-import kitchenpos.domain.menu.repository.MenuRepository;
-import kitchenpos.domain.order.repository.OrderRepository;
-import kitchenpos.domain.ordertable.repository.OrderTableRepository;
-import kitchenpos.domain.product.repository.ProductRepository;
+import kitchenpos.domain.menugroup.MenuGroupRepository;
+import kitchenpos.domain.menu.MenuRepository;
+import kitchenpos.domain.order.OrderRepository;
+import kitchenpos.domain.ordertable.OrderTableRepository;
+import kitchenpos.domain.product.ProductRepository;
 import kitchenpos.ui.dto.request.CreateOrderLineItemRequest;
 import kitchenpos.ui.dto.request.CreateOrderRequest;
 import kitchenpos.ui.dto.request.UpdateOrderStatusRequest;
