+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -11,7 +11,7 @@ import kitchenpos.application.dto.ReadProductDto;
 import kitchenpos.config.IntegrationTest;
 import kitchenpos.domain.product.Product;
 import kitchenpos.domain.exception.InvalidPriceException;
-import kitchenpos.domain.product.repository.ProductRepository;
+import kitchenpos.domain.product.ProductRepository;
 import kitchenpos.ui.dto.request.CreateProductRequest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
