+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -12,13 +12,13 @@ import kitchenpos.application.dto.ReadOrderTableDto;
 import kitchenpos.application.exception.OrderTableNotFoundException;
 import kitchenpos.config.IntegrationTest;
 import kitchenpos.domain.exception.InvalidNumberOfGuestsException;
-import kitchenpos.domain.menu.repository.MenuRepository;
-import kitchenpos.domain.menugroup.repository.MenuGroupRepository;
-import kitchenpos.domain.order.repository.OrderRepository;
+import kitchenpos.domain.menu.MenuRepository;
+import kitchenpos.domain.menugroup.MenuGroupRepository;
+import kitchenpos.domain.order.OrderRepository;
 import kitchenpos.domain.ordertable.OrderTable;
-import kitchenpos.domain.ordertable.repository.OrderTableRepository;
-import kitchenpos.domain.product.repository.ProductRepository;
-import kitchenpos.domain.tablegroup.repository.TableGroupRepository;
+import kitchenpos.domain.ordertable.OrderTableRepository;
+import kitchenpos.domain.product.ProductRepository;
+import kitchenpos.domain.tablegroup.TableGroupRepository;
 import kitchenpos.ui.dto.request.CreateOrderTableRequest;
 import kitchenpos.ui.dto.request.UpdateOrderTableEmptyRequest;
 import kitchenpos.ui.dto.request.UpdateOrderTableNumberOfGuestsRequest;
