+++ src/test/java/kitchenpos/domain/ordertable/GroupingOrderTableServiceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.ordertable.service;
+package kitchenpos.domain.ordertable;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -16,20 +16,18 @@ import kitchenpos.domain.exception.InvalidOrderTableStatusException;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menu.MenuProduct;
 import kitchenpos.domain.menu.MenuProducts;
-import kitchenpos.domain.menu.repository.MenuRepository;
+import kitchenpos.domain.menu.MenuRepository;
 import kitchenpos.domain.menugroup.MenuGroup;
-import kitchenpos.domain.menugroup.repository.MenuGroupRepository;
+import kitchenpos.domain.menugroup.MenuGroupRepository;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItem;
+import kitchenpos.domain.order.OrderRepository;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.order.repository.OrderRepository;
-import kitchenpos.domain.ordertable.OrderTable;
-import kitchenpos.domain.ordertable.repository.OrderTableRepository;
 import kitchenpos.domain.product.Product;
-import kitchenpos.domain.product.repository.ProductRepository;
+import kitchenpos.domain.product.ProductRepository;
+import kitchenpos.domain.tablegroup.GroupingTableService;
 import kitchenpos.domain.tablegroup.TableGroup;
-import kitchenpos.domain.tablegroup.repository.TableGroupRepository;
-import kitchenpos.domain.tablegroup.service.GroupingTableService;
+import kitchenpos.domain.tablegroup.TableGroupRepository;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
